Clean up recommendation code, add comments
