+++ src/server/daemons/GenerateRecommendationDaemon.java
@@ -104,20 +104,6 @@ public class GenerateRecommendationDaemon {
 		return correlation != null ? correlation.getCorrelationRatio(userPref.getPopularity()) : 0.0;
 	}
 	
-
-	/**
-	 * Given a map of preferences to correlation scores, finds the entry in the map with the highest correlation score.  This entry, if it
-	 *   exists, will provide the user recommendation.
-	 * 
-	 * @param correlationScoreMap Map of preferences to their associated correlation scores (with all user preferences).
-	 * @param comparator Used to order the entries of the map by correlation score
-	 * @return The map entry with the maximum correlation score or Optional.empty() if there are no preferences that are correlated with user preferences
-	 */
-	private Optional<Map.Entry<Preference, Double>> getMaxMapEntry(
-			Map<Preference, Double> correlationScoreMap,
-			Comparator<Map.Entry<Preference, Double>> comparator) {
-		return correlationScoreMap.isEmpty() ? Optional.empty() : Optional.of(Collections.max(correlationScoreMap.entrySet(), comparator));
-	}
 	
 	/**
 	 * Find the best preference recommendation, based on correlation score, relative to user preferences.
@@ -156,205 +142,4 @@ public class GenerateRecommendationDaemon {
 		// Finally create a recommendation object out of the map entry with the greatest correlation score
 		return topScoredEntry.map(tuple -> new Recommendation(tuple._1(), user, tuple._2()));
 	}
-	
-	/*
-	private double getNewCorrelationScore(Double originalScore, double originalWeight, double newWeight, double originalPopularity, double newPopularity) {
-		return (originalScore == null) ? newWeight / newPopularity :
-			(newWeight/newPopularity - originalWeight/originalPopularity) + originalScore;
-	}
-	
-	Map<Preference, Double> getUpdatedPreferenceScores(UserProfile user,
-			PreferenceCorrelationGraph correlationGraph,
-			List<UpdatePreferenceRequest> preferenceUpdates,
-			PreferenceCategory preferenceCategory,
-			Map<Preference, Double> originalCorrelationScores) {
-		@SuppressWarnings("serial")
-		Map<Preference, Double> newCorrelationScores = new HashMap<Preference, Double>() {{
-			putAll(originalCorrelationScores);
-		}};
-		
-		for (UpdatePreferenceRequest update : preferenceUpdates) {
-			handlePreferenceUpdate(user, preferenceCategory,
-					newCorrelationScores, update);	
-		}
-		
-		return newCorrelationScores;
-	}
-
-	private void handlePreferenceUpdate(UserProfile user,
-			PreferenceCategory preferenceCategory,
-			Map<Preference, Double> correlationScores,
-			UpdatePreferenceRequest update) {
-		Preference targetPref = update.getPreferenceToUpdate();
-		
-		// For each correlation update, match up preference to UpdateAction
-		Map<Preference, UpdateAction> correlationUpdates = update.getCorrelationUpdates().entrySet().stream()
-			       .collect(Collectors.toMap(entry -> entry.getKey().getToPreference(), 
-			    		   entry -> entry.getValue()));
-			
-		Set<PreferenceCorrelation> allCorrelations = update.getPreferenceToUpdate().getCorrelations();
-		
-		for (PreferenceCorrelation prefCorrelation : allCorrelations) {
-			Preference correlatedPreference = prefCorrelation.getToPreference();
-		
-			double newWeight = prefCorrelation.getWeight();
-			
-			double originalWeight = getOriginalCorrelationWeight(correlationUpdates,
-					correlatedPreference, newWeight);
-				
-			if (isUserPreference(user, preferenceCategory, targetPref)) {
-				updateCorrelationScoresBasedOnUserPreferenceUpdate(user,
-						preferenceCategory, correlationScores, update,
-						targetPref, correlatedPreference, newWeight,
-						originalWeight);
-
-			} else if (isUserPreference(user, preferenceCategory, correlatedPreference) &&
-					correlationUpdates.containsKey(correlatedPreference)) {
-				updateCorrelationScoresBasedOnCorrelatedPreferenceUpdate(
-						correlationScores, targetPref, correlatedPreference,
-						newWeight, originalWeight);
-		   }
-		}
-	}
-
-	private void updateCorrelationScoresBasedOnUserPreferenceUpdate(
-			UserProfile user, PreferenceCategory preferenceCategory,
-			Map<Preference, Double> correlationScores,
-			UpdatePreferenceRequest update, Preference targetPref,
-			Preference correlatedPreference, double newWeight,
-			double originalWeight) {
-		double newPopularity = targetPref.getPopularity();
-		
-		double originalPopularity = newPopularity - update.getPopularityUpdate().getDelta();
-		
-		if (! isUserPreference(user, preferenceCategory, correlatedPreference)) {
-			correlationScores.compute(correlatedPreference, 
-				(key, value) ->
-		             getNewCorrelationScore(value, originalWeight, newWeight, originalPopularity, newPopularity));
-		}
-	}
-
-	private double getOriginalCorrelationWeight(
-			Map<Preference, UpdateAction> correlationUpdates,
-			Preference correlatedPreference, double newWeight) {
-		return correlationUpdates.containsKey(correlatedPreference) ?
-				newWeight - correlationUpdates.get(correlatedPreference).getDelta() :
-					newWeight;
-	}
-
-	private void updateCorrelationScoresBasedOnCorrelatedPreferenceUpdate(
-			Map<Preference, Double> correlationScores, Preference targetPref,
-			Preference correlatedPreference, double newWeight,
-			double originalWeight) {
-		double originalPopularity = correlatedPreference.getPopularity();
-		
-		correlationScores.compute(targetPref, 
-				(key, value) -> getNewCorrelationScore(value, originalWeight, newWeight, originalPopularity, originalPopularity));
-	}
-
-	private boolean isUserPreference(UserProfile user,
-			PreferenceCategory preferenceCategory, Preference targetPref) {
-		return user.getPreferencesForCategory(preferenceCategory).contains(targetPref);
-	}
-	*/
-	
-	/*
-	if (isUserPreference(user, preferenceCategory, targetPref)) {
-		double newPopularity = originalPopularity + update.getPopularityUpdate().getDelta();
-		
-		for (PreferenceCorrelation prefCorrelation : update.getPreferenceToUpdate().getCorrelations()) {
-			Preference correlatedPreference = prefCorrelation.getToPreference();
-			double originalWeight = prefCorrelation.getWeight();
-			double newWeight = correlationUpdates.containsKey(correlatedPreference) ?
-					originalWeight + correlationUpdates.get(correlatedPreference).getDelta() :
-						originalWeight;
-			newCorrelationScores.compute(correlatedPreference, 
-					(key, value) ->
-			             getNewCorrelationScore(value, originalWeight, newWeight, originalPopularity, newPopularity));
-		}
-	} else {				
-		for (PreferenceCorrelation prefCorrelation : targetPref.getCorrelations()) {
-			Preference correlatedPref = prefCorrelation.getToPreference();
-			if (isUserPreference(user, preferenceCategory, correlatedPref) &&
-					correlationUpdates.containsKey(correlatedPref)) {
-				double originalWeight = prefCorrelation.getWeight();
-				double newWeight = correlationUpdates.get(correlatedPref).getDelta() + originalWeight;
-			
-				newCorrelationScores.compute(targetPref, 
-						(key, value) -> getNewCorrelationScore(value, originalWeight, newWeight, originalPopularity, originalPopularity));
-			}
-		}	
-	}
-	*/
-	
-	/*
-	Map<Preference, Double> getUpdatedPreferenceScores(UserProfile user,
-			PreferenceCorrelationGraph correlationGraph,
-			List<UpdatePreferenceRequest> preferenceUpdates,
-			PreferenceCategory preferenceCategory,
-			Map<Preference, Double> originalCorrelationScores) {
-		Stream<Tuple2<Preference, Map.Entry<PreferenceCorrelation, UpdateAction>>> prefUpdatesStream =
-		             preferenceUpdates.stream()
-			.filter(upr -> originalCorrelationScores.containsKey(upr.getPreferenceToUpdate()))
-			.filter(upr -> upr.getPopularityUpdate() != null)
-			.flatMap(upr -> upr.getCorrelationUpdates().entrySet().stream().map(cu -> 
-			    new Tuple2<Preference, Map.Entry<PreferenceCorrelation, UpdateAction>>(
-					upr.getPreferenceToUpdate(), cu)));
-			//.map((Tuple2<Preference, Map.Entry<PreferenceCorrelation, UpdateAction>> tuple) -> tuple._1() );
-		
-		Set<Preference> userPreferences = user.getPreferences().get(preferenceCategory);
-		Map<Preference, Double> preferenceUpdateScores = prefUpdatesStream
-		    .filter(cu -> userPreferences.contains(cu._2().getKey().getToPreference()))
-		    .map(cu -> new Tuple2<Preference, Double>(cu._1(), (double) cu._2().getValue().getDelta() / (double) correlationGraph.getPreference(cu._2().getKey().getToPreference().getID(), preferenceCategory).getPopularity()))
-		    .collect(Collectors.groupingBy((Tuple2<Preference, Double> tuple) -> tuple._1(),
-                                                   Collectors.summingDouble((Tuple2<Preference, Double> tuple) -> tuple._2())));
-		
-		Map<Preference, Double> userPrefPopularityUpdates = preferenceUpdates.stream()
-		    .filter(upr -> user.getPreferencesForCategory(preferenceCategory).contains(upr.getPreferenceToUpdate()))
-		    .filter(upr -> upr.getPopularityUpdate() != null)
-		    .collect(Collectors.groupingBy((UpdatePreferenceRequest upr) -> upr.getPreferenceToUpdate(),
-		    		Collectors.summingDouble((UpdatePreferenceRequest upr) -> (double) upr.getPopularityUpdate().getDelta())));
-		
-		Map<Preference, Double> correlationScoreChange = new HashMap<Preference, Double>();
-		for (UpdatePreferenceRequest upr : preferenceUpdates) {
-			if (! user.getPreferencesForCategory(preferenceCategory).contains(upr.getPreferenceToUpdate())) {
-				continue;
-			}
-			
-			if (upr.getPopularityUpdate() == null) {
-				continue;
-			}
-			
-			UpdateAction popularityUpdate = upr.getPopularityUpdate();
-			
-			Map<PreferenceCorrelation, UpdateAction> correlationUpdates = upr.getCorrelationUpdates();
-		    for (PreferenceCorrelation correlation : upr.getPreferenceToUpdate().getCorrelations()) {
-		    	int newPopularity = upr.getPreferenceToUpdate().getPopularity();
-		    	int oldPopularity = newPopularity - popularityUpdate.getDelta();
-		    	UpdateAction correlationUpdate = correlationUpdates.get(correlation);
-		    	int newWeight = correlation.getWeight();
-		    	int oldWeight = newWeight - ((correlationUpdate != null) ? correlationUpdate.getDelta() : 0);
-		    	
-		    	double currScoreChange = 0.0;
-		    	if (correlationScoreChange.containsKey(correlation)) {
-		    		currScoreChange = correlationScoreChange.get(upr.getPreferenceToUpdate());
-		    	}
-		    	
-		    	currScoreChange += (newWeight / newPopularity) - (oldWeight / oldPopularity);
-		    	correlationScoreChange.put(upr.getPreferenceToUpdate(), currScoreChange);
-		    }
-		}
-		
-		Map<Preference, Double> correlationScores =  new HashMap<Preference, Double>();
-		    
-		for (Preference pref : originalCorrelationScores.keySet()) {
-			double originalScore = originalCorrelationScores.get(pref);
-			Double userPrefChange = correlationScoreChange.containsKey(pref) ? correlationScoreChange.get(pref) : 0.0;
-			Double prefUpdate = preferenceUpdateScores.containsKey(pref) ? preferenceUpdateScores.get(pref) : 0.0;
-			correlationScores.put(pref, originalScore + userPrefChange + prefUpdate);
-		}
-		
-		return correlationScores;
-	}
-	*/
 }
